(PE-26709) Restart Solaris 11 service on upgrade
Prior to this commit, the `refresh` action was executed by SMF when
the package manifest changed. However, no action was implemented for
`refresh`, hence the mcollective service did not automatically restart
when the puppet-agent package was upgraded.

This commit adds functionality to the `refresh` action, killing the
service for it to be restarted automatically by SMF, which is the
intended behavior.